move tag services into one tag.service
