+++ src/app/api/tags/task-template/route.ts
@@ -1,13 +1,13 @@
 import { type NextRequest } from 'next/server';
 import { handleError } from '@/lib/utils/errors.utils';
-import TaskTemplateTagService from '@/lib/services/task-template-tag.service';
+import TagService from '@/lib/services/tag.service';
 import { createTagSchema } from '@/lib/schemas/tag.schema';
 import { getSession } from '@/lib/utils/auth.utils';
 
 export async function GET() {
     try {
         const session = await getSession();
-        const tags = await TaskTemplateTagService.getTaskTemplateTagsByOrgId(
+        const tags = await TagService.getTaskTemplateTagsByOrgId(
             session.activeOrganizationId
         );
         return Response.json({ data: tags }, { status: 200 });
@@ -22,7 +22,7 @@ export async function POST(request: NextRequest) {
         const body = await request.json();
         const parsed = createTagSchema.parse(body);
 
-        const tag = await TaskTemplateTagService.createTaskTemplateTag(
+        const tag = await TagService.createTaskTemplateTag(
             parsed.name,
             session.activeOrganizationId,
             parsed.colorHexCode
